requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
dj-database-url==2.2.0
Django==5.1.1
django-redis==5.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.1.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stripe==11.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.7.0
django-cors-headers==4.6.0
channels==4.0.0
daphne==4.1.0
